requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.5
pandas>=2.0.0
numpy>=1.23.0
python-dotenv>=1.0.0
requests>=2.26.0
httpx>=0.24.0
tenacity>=8.2.2
google-generativeai>=0.3.0